Doc: Documented supported tool versions
Add a section on supported tools to the Ibex documentation. Populate it
with my knowledge about tools we've tried. That's a starting point: we
need to add more version numbers and other information as we collect
more data points.

The Verilator version is taken from the new tool_requirements.py file
and inserted into Sphinx.