[rtl] Forward register data with status
Instead of accessing the signals via the module instance, use the
signals connected to the output port of the module.

Only set the values for RS1/2 if they are used.

Without this addition an instruction without a valid encoding for a
register would reuse invalid data as the address of the register.
Certain checks require that the data must match the register content if
the address is non-zero.

Reuse the signal from the instruction decoder to set the registers to
non-zero values only if the instruction contains a valid encoding for
the register.